// src/modules/Client/ClientModal/client.modal.types.ts
export type formValues = {
  clientId:  number;
  name: string;
  firstSurname: string;
  secondSurname: string;
  cedula: string;
  age: number;
  phoneNumber: string;
  nationality: string;
  observation: string;
  hasAnyDisability: boolean;
  disability: string;
  isAlergic: boolean;
  alergic: string;
  isLactoseIntolerant: boolean;
  creditCardNumber: string;
  cvc: string;
  expiredDate: string;
  cardName: string;
}